% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn003.R
\name{testtest}
\alias{testtest}
\title{A ml modular framework}
\arguments{
\item{testtest}{type stuff}
}
\description{
This function allows you to express your love of cats.
}
\examples{
rnetf()
}
\keyword{testtest}
